lesson-7: refactor dagger modules
