+++ app/src/main/java/com/rogoz208/mygithubclient/App.kt
@@ -2,11 +2,11 @@ package com.rogoz208.mygithubclient
 
 import android.app.Application
 import android.content.Context
-import com.rogoz208.mygithubclient.di.dagger.DaggerMyComponent
-import com.rogoz208.mygithubclient.di.dagger.MyComponent
+import com.rogoz208.mygithubclient.di.dagger.AppComponent
+import com.rogoz208.mygithubclient.di.dagger.DaggerAppComponent
 
 class App : Application() {
-    val di: MyComponent by lazy { DaggerMyComponent.builder().build() }
+    val di: AppComponent by lazy { DaggerAppComponent.builder().build() }
 }
 
 val Context.app: App
